fix: derive care-log draft participant from current user
Replace the hardcoded diary draft participant with the logged-in user so create flows show the same identity that will actually be sent to the backend.

Constraint: Draft entries still need a display-ready participant object before the backend returns a persisted care-log item
Rejected: Keep the hardcoded placeholder participant | mismatches the authenticated user and confuses create-form rendering
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: If avatar or participant mapping is shared across more pages, extract a common mapper instead of reintroducing page-local placeholders
Tested: npm run typecheck; npx eslint src/pages/CareLog/utils/createDraftCareLogEntry.ts src/pages/CareLog/CalendarPage.tsx src/pages/CareLog/CareLogCreatePage.tsx
Not-tested: Browser verification of draft participant rendering before and after login state changes
