+++ src/pages/CareLog/CalendarPage.tsx
@@ -21,27 +21,9 @@ import CareLogModal, {
 import useCreateCareLogEntry from '@/pages/CareLog/hooks/useCreateCareLogEntry';
 import useGetCareLogEntries from '@/pages/CareLog/hooks/useGetCareLogEntries';
 import type { CareLogEntry } from '@/pages/CareLog/types';
+import createDraftCareLogEntry from '@/pages/CareLog/utils/createDraftCareLogEntry';
 
 const defaultSelectedDate = new Date();
-
-const defaultParticipant = {
-  id: 'current-user',
-  name: '王希銘',
-  src: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEiqrpYjz-y8bMs_qvQFR_w4vW_HEUAsQwzgMSbzLMJFytcdMUrY4M25Jx7EjoGDbvSIRaagzEacgR2hIhCLy39aMqWGH9cR-MQ3LjZzljWWCoDjzgU2y7G9nisZk47dRYesEYrG9Bg79XhA/s400/nigaoe_nakajima_atsushi.png',
-};
-
-function createDraftCareLogEntry(selectedDate = new Date()): CareLogEntry {
-  return {
-    id: globalThis.crypto?.randomUUID?.() ?? `diary-${Date.now()}`,
-    title: '',
-    description: '',
-    startsAt: format(selectedDate, "yyyy-MM-dd'T'HH:mm:ssxxx"),
-    repeatPattern: 'none',
-    participants: [defaultParticipant],
-    status: 'pending',
-    isImportant: false,
-  };
-}
 
 function getSelectedDateFromState(state: unknown) {
   if (!state || typeof state !== 'object' || !('selectedDate' in state)) {
